+++ qis/plots/boxplot.py
@@ -509,4 +509,3 @@ def df_dict_boxplot_by_classification_var(data_dict: Dict[Tuple[str, str], pd.Da
                         textcoords='offset points', ha='left', va='bottom',
                         fontsize=12)
             ax.scatter(x=last[x_hue_name], y=last[key[1]], s=20)
-0